backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
Django==4.0.5
django-cors-headers==3.13.0
django-mysql==4.7.1
djangorestframework==3.13.1
djangorestframework-queryfields==1.0.0
docopt==0.6.2
idna==3.3
jsonfield==3.1.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
pathspec==0.9.0
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==2.5.2
protobuf==3.20.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.9
XlsxWriter==3.0.3
yarg==0.1.9